[pay]fix pingppviews

+++ shopmanager/flashsale/pay/views.py
@@ -105,7 +105,7 @@ class PINGPPChargeView(View):
             
             sku = ProductSku.objects.get(pk=form.get('sku_id'),product=product)
             discount_fee = sku.calc_discount_fee(xlmm=xlmm)
-            real_fee = int(sku.agent_price * int(form.get('num')) * 100 - discount_fee * 100)
+            real_fee = int(sku.agent_price * int(form.get('num')) * 100) - int(discount_fee * 100)
             
             assert payment > 0 and payment == real_fee ,u'订单金额有误'
